grade-simulator-service/dummy_data/requirements.txt
flask==2.0.1
flask-restful==0.3.9
sqlalchemy==1.4.23
python-dotenv==0.19.0
flask-cors==3.0.10
